# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FAA NOTAM: Developer Portal API",
    "description": "Backend REST API for the NOTAM Developer Portal",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhot:8080/api/ver{version}/",
      "description": "Local Development",
      "variables": {
        "version": {
          "default": "1",
          "description": "The API version number"
        }
      }
    }
  ],
  "paths": {
...
    "/user/{id}": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Deletes user with the given Id.",
        "description": "<p>Allowed roles: Admin</p>\r\n<p>Only admin users can be deleted. Return 403 Status Code if Admin tries to delete regular user.</p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The user Id.",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Retrieves User with the given Id.",
        "description": "<p>Allowed roles: Admin and Regular User</p>\r\n<p>Regular User can only access his/her own account</p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The User Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The User Id.",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserModel"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": [ ]
          }
        ]
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Updates user with the given Id.",
        "description": "<p>Allowed roles: Admin and Regular User can update self accounts only</p>\r\n<p>Validate and save password similar to Register endpoint.</p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The user Id.",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "The update user model.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": [ ]
          }
        ]
      }
    },
    "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Searches users matching given criteria parameters.",
        "description": "<p>Allowed roles: Admin</p>\r\n<p>\r\n            Provided parameters are just an example. Final parameter list will be defined during development phase based on the wireframes. (same for returned user fields)\r\n            </p>\r\n<p>All string fields will be matched using 'Contains' comparison. All non-null criterias should be AND-ed.</p>\r\n<p>Default and maximum 'limit' should be configurable in backend config.</p>",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "The user name criteria.",
            "schema": {
              "type": "string",
              "description": "The user name criteria.",
              "nullable": true
            }
          },
          {
            "name": "email",
            "in": "query",
            "description": "The email criteria.",
            "schema": {
              "type": "string",
              "description": "The email criteria.",
              "nullable": true
            }
          },
          {
            "name": "isApproved",
            "in": "query",
            "description": "The flag criteria indicating whether to search approved/denied users.",
            "schema": {
              "type": "boolean",
              "description": "The flag criteria indicating whether to search approved/denied users.",
              "nullable": true
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "The name of the field to sort by.",
            "schema": {
              "type": "string",
              "description": "The name of the field to sort by.",
              "default": "Id",
              "nullable": true
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "description": "The sort order.",
            "schema": {
              "$ref": "#/components/schemas/SortOrder"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "The offset.",
            "schema": {
              "type": "integer",
              "description": "The offset.",
              "format": "int32",
              "default": 0,
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The limit.",
            "schema": {
              "type": "integer",
              "description": "The limit.",
              "format": "int32",
              "default": 20,
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserModelSearchResult"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": [ ]
          }
        ]
      }
    },
    "/user/register/apiUser": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Registers new regular user.",
        "description": "<p>reCaptcha verification: Required. To validate use reCAPTCHA secret key and 'ReCaptchaResponse' from the request model. The reCAPTCHA secret key must be configured in backend config.\r\n            For more details, see - https://developers.google.com/recaptcha/docs/verify\r\n            </p>\r\n<p>Password strength should be configurable in backend config. Return Bad Request in case password validation fails.</p>\r\n<p>Generate random email verification code (Base64URL). Construct URL using 'userId' and 'verification code' values. Send email to the user with constructed URL. The URL template should be configurable in backend config.</p>\r\n<p>Email must be unique. If user with same Email is already registered, then if IsApproved!=NULL return 409 (Conflict) Status Code;\r\n            otherwise update fields in existing record and proceed with the standard registration flow - this way we are allowing updating data and re-sending verification email.</p>",
        "requestBody": {
          "description": "The register model.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully registered"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT - if there is some conflict between provided and existing data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        }
      }
    },
    "/user/register/admin": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Registers new admin user.",
        "description": "<p>Allowed roles: Admin</p>\r\n<p>Re-Captcha verification: not needed, because it is not anonymous operation (Admin already logged in).</p>\r\n<p>Password strength should be configurable in backend config. Return Bad Request in case password validation fails.</p>\r\n<p>Generate random email verification code (Base64URL). Construct URL using 'userId' and 'verification code' values. Send email to the user with constructed URL. The URL template should be configurable in backend config.</p>",
        "requestBody": {
          "description": "The register model.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAdminModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAdminModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAdminModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedItemModel"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": [ ]
          }
        ]
      }
    },
    "/user/{id}/verifyEmail": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Verifies user email.",
        "description": "<p>\r\n            Compare the provided code with the 'EmailConfirmationCode' of the corresponding user.\r\n            If code matches, set the 'IsEmailConfirmed'=True. If user is Admin, set 'IsEnabled'=True and 'IsApproved'=True (no need to approve admin user).\r\n            </p>\r\n<p>If matching combination of user and code is not found, return 404 Status Code.</p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The user Id.",
              "format": "int64"
            }
          },
          {
            "name": "code",
            "in": "query",
            "description": "The email verification code.",
            "schema": {
              "type": "string",
              "description": "The email verification code.",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        }
      }
    },
    "/user/{id}/approve": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Approves user registration.",
        "description": "<p>Allowed roles: Admin</p>\r\n<p>\r\n            If 'IsApproved' for the given user is not NULL, return 409 Status Code - don't allow to Approve/Reject multiple times.\r\n            </p>\r\n<p>\r\n            Set 'IsEnabled'=True and 'IsApproved'=True.\r\n            </p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The user Id.",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT - if there is some conflict between provided and existing data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": [ ]
          }
        ]
      }
    },
    "/user/{id}/deny": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Denies user registration.",
        "description": "<p>Allowed roles: Admin</p>\r\n<p>\r\n            If 'IsApproved' for the given user is not NULL, return 409 Status Code - don't allow to Approve/Reject multiple times.\r\n            </p>\r\n<p>\r\n            Set 'IsApproved'=False.\r\n            </p>",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user Id.",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "The user Id.",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "BAD REQUEST - if there was problem with the request (e.g. malformed or some parameters are missing).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "NOT AUTHORIZED - if the request didn't bear authentication information or the authentication information is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN - if authentication was successful, but access is denied for the user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND - if some entity with the given Id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT - if there is some conflict between provided and existing data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL SERVER ERROR - if the request was properly formatted, but the operation failed on the server side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorModel": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Gets or sets the message.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "An API error Model."
      },
...
      "UpdateUserModel": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "The full name of the user.",
            "nullable": true,
            "example": "John Doe"
          },
          "password": {
            "type": "string",
            "description": "The user password.",
            "nullable": true,
            "example": "Secret!23"
          },
          "company": {
            "type": "string",
            "description": "The user's Company.",
            "nullable": true,
            "example": "Topcoder"
          },
          "address": {
            "type": "string",
            "description": "The user's Address.",
            "nullable": true,
            "example": "Main Street, 123"
          },
          "city": {
            "type": "string",
            "description": "The user's City.",
            "nullable": true,
            "example": "Ney York"
          },
          "state": {
            "type": "string",
            "description": "The user's State.",
            "nullable": true,
            "example": "NY"
          },
          "country": {
            "type": "string",
            "description": "The user's Country.",
            "nullable": true,
            "example": "United States"
          },
          "zipCode": {
            "type": "string",
            "description": "The Zip Code.",
            "nullable": true,
            "example": "12345"
          },
          "primaryPhone": {
            "type": "string",
            "description": "The user's primary phone number.",
            "nullable": true,
            "example": "+1-541-754-3010"
          },
          "alternatePhone": {
            "type": "string",
            "description": "The user's alternate phone number.",
            "nullable": true,
            "example": "+1-754-541-1030"
          },
          "notamDataIntendedUsage": {
            "type": "string",
            "description": "The intended NOTAM data usage.",
            "nullable": true,
            "example": "Private pilot flights in US."
          }
        },
        "additionalProperties": false,
        "description": "The Update user model."
      },
...
    },
    "securitySchemes": {
      "jwt_bearer": {
        "type": "http",
        "description": "Authentication using JWT Token in header.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}